laravel 8.22 fix

+++ app/Http/Controllers/DashboardPostController.php
@@ -7,6 +7,7 @@ use App\Models\Category;
 use Illuminate\Http\Request;
 use Illuminate\Support\Str;
 use \Cviebrock\EloquentSluggable\Services\SlugService;
+use Illuminate\Support\Facades\Storage;
 
 class DashboardPostController extends Controller
 {
@@ -116,13 +117,22 @@ class DashboardPostController extends Controller
             'title' => 'required|max:255',
             'slug' => 'required|unique:posts',
             'category_id' => 'required',
+             'image' => 'image|file|max:1024', // maksimal ukuran file
             'body' => 'required'
         ];
+        
         // Jika slug tidak berubah, maka tidak perlu validasi unik
         if ($request->slug != $post->slug) {
             $rules['slug'] = 'required|unique:posts';
         }
         $validatedData = $request->validate($rules);
+
+         if ($request->file('image')){
+            if($request->oldImage){
+                Storage::delete($request->oldImage);
+            }
+            $validatedData['image'] = $request->file('image')->store('post-images'); // simpan file gambar ke folder post-images
+        }
          $validatedData['user_id'] = auth()->user()->id;
         $validatedData['excerpt'] = Str::limit(strip_tags($request->body), 200);
 
@@ -140,6 +150,9 @@ class DashboardPostController extends Controller
      */
     public function destroy(Post $post)
     {
+        if($post->image){
+                Storage::delete($post->image);
+            }
          Post::destroy($post->id); // simpan data post ke database
         // Redirect ke halaman posts dengan pesan sukses
